more restructuring wip
