+++ ZipUtility/Logging/Structured/StructuredLoggerExtensions.cs
@@ -7,10 +7,10 @@ using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace SRF.FileLogging
+namespace SRF.FileLogging.Structured
 {
   /// <summary>File logger extension methods.</summary>
-  static public class FileLoggerExtensions
+  static public class StructuredLoggerExtensions
   {
     /// <summary>
     /// Adds the file logger provider, aliased as 'File', in the available services as singleton and binds the file logger options class to the 'File' section of the appsettings.json file.
@@ -19,16 +19,16 @@ namespace SRF.FileLogging
     {
       builder.AddConfiguration();
 
-      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, FileLoggerProvider>());
-      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IConfigureOptions<FileLoggerOptions>, FileLoggerOptionsSetup>());
-      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IOptionsChangeTokenSource<FileLoggerOptions>, LoggerProviderOptionsChangeTokenSource<FileLoggerOptions, FileLoggerProvider>>());
+      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, StructuredLoggerProvider>());
+      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IConfigureOptions<StructuredLoggerOptions>, StructuredLoggerOptionsSetup>());
+      builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IOptionsChangeTokenSource<StructuredLoggerOptions>, LoggerProviderOptionsChangeTokenSource<StructuredLoggerOptions, StructuredLoggerProvider>>());
       return builder;
     }
 
     /// <summary>
     /// Adds the file logger provider, aliased as 'File', in the available services as singleton and binds the file logger options class to the 'File' section of the appsettings.json file.
     /// </summary>
-    static public ILoggingBuilder AddFileLogger(this ILoggingBuilder builder, Action<FileLoggerOptions> configure)
+    static public ILoggingBuilder AddFileLogger(this ILoggingBuilder builder, Action<StructuredLoggerOptions> configure)
     {
       if (configure == null)
       {
